Update H-6.tex
The original file, in the first question, had a duplicate item of Distance Metrics. I do not believe that there should be two versions of the same item in a single question, and I am guessing it was a miss type. Either it needs to be a new question or it was mistakenly written twice, and a new question does not need to exist.

My pull request is for H-6 to not have a secondary Distance Metrics item on the first question, or to have a different value for that item space. I removed the secondary copy of Distance Metrics, and will state that if it stayed, then you would have the exact answer twice in the same question.